WIP major refactor
